godeps: maybebtc renamed is account
