+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/.travis.yml
