+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/entry_test.go
