+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/examples/basic/basic.go
