+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/examples/hook/hook.go
