+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/formatter.go
