+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/formatter_bench_test.go
