+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/hook_test.go
