+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/hooks/airbrake/airbrake.go
