+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/hooks/papertrail/README.md
