+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/hooks/syslog/README.md
