+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/hooks/syslog/syslog.go
