+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/json_formatter.go
