+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/logrus.go
