+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/logrus_test.go
