+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/terminal_notwindows.go
