+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/terminal_windows.go
