+++ Godeps/_workspace/src/github.com/briantigerchow/logrus/text_formatter.go
