+++ Godeps/_workspace/src/github.com/briantigerchow/pubsub/README.md
@@ -6,7 +6,7 @@ View the [API Documentation](http://godoc.org/github.com/tuxychandru/pubsub).
 
 ## License
 
-Copyright (c) 2013, Chandra Sekar S
+Copyright (c) 2013, Chandra Sekar S  
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
