--- Godeps/_workspace/src/github.com/maybebtc/pubsub/Godeps/Godeps.json
@@ -1,13 +1,0 @@
-{
-	"ImportPath": "github.com/maybebtc/pubsub",
-	"GoVersion": "go1.3.3",
-	"Packages": [
-		"./..."
-	],
-	"Deps": [
-		{
-			"ImportPath": "gopkg.in/check.v1",
-			"Rev": "64131543e7896d5bcc6bd5a76287eb75ea96c673"
-		}
-	]
-}
