--- Godeps/_workspace/src/github.com/maybebtc/pubsub/Godeps/Readme
@@ -1,5 +1,0 @@
-This directory tree is generated automatically by godep.
-
-Please do not edit.
-
-See https://github.com/tools/godep for more information.
