--- Godeps/_workspace/src/github.com/maybebtc/pubsub/Makefile
@@ -1,2 +1,0 @@
-vendor:
-	godep save -r ./...
